requirements.txt
psycopg2-binary
pandas
sqlalchemy
flask
flask_restx
logging
pytest